moved sendEmail function to helpers.js
