--- lib/data-helpers.js
@@ -1,9 +1,0 @@
-const md5 = require('md5')
-
-const urlCreator = () => {
-  const string = 'where should we go for lunch?'
-  const starter = Math.random().toString(36).substring(7);
-  return md5(starter+string)
-}
-
-module.exports = { urlCreator };
